requirements.txt

flask==2.3.3
chromadb==0.4.15
openai==1.3.0
python-dotenv==1.0.0
PyPDF2==3.0.1
docx2txt==0.8
numpy==1.26.4
scikit-learn==1.5.2
tiktoken
pdfplumber==0.9.0
requests==2.32.3
gunicorn==21.2.0
tqdm==4.66.4
pandas==2.2.3
